Create uighur_arabic.yml
A table for the Uighur language using the Arabic script, based on the MS Word version (sent to me by Jessalyn Zoom) of this table: https://www.loc.gov/catdir/cpso/romanization/uighur.pdf.  The mappings in the table incorporate points 1, 2, and 3 under "Rules of Application", but no specific words and examples (i.e., the table only captures the rules, not the exceptions, if any).