Update uighur_arabic.yml
I was able to consult with a Uighur specialist, and now have a better understanding of this script.  In this new version of the table, the r2s mapping only uses Arabic characters in the U+06XX block.  This has simplified this section, as these mappings are not context-dependent except for initial vowels.  The compatibility characters in the U+FEXX range are still included in the s2r section (alongside the U+06XX versions), though in practice they will not  be encountered often.

I have some test phrases that I can provide as well - please let me know where and in what format I should submit these.